+++ crates/renzora_runtime/src/plugins.rs
@@ -26,6 +26,7 @@ pub(crate) fn add_runtime_plugins(app: &mut App) {
     app.add_plugins(renzora_atmosphere::AtmospherePlugin::default());
     #[cfg(feature = "audio")]
     app.add_plugins(renzora_audio::AudioPlugin::default());
+    app.add_plugins(renzora_auto_exposure::AutoExposurePlugin::default());
     #[cfg(feature = "bloom")]
     app.add_plugins(renzora_bloom_effect::BloomEffectPlugin::default());
     #[cfg(feature = "blueprint")]